backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
psycopg2-binary==2.9.10
